added quality check to trino snapshot and insert template
