+++ projects/vdk-plugins/vdk-trino/src/vdk/plugin/trino/templates/load/dimension/scd1/04-create-tmp-target.sql
@@ -1,0 +1,3 @@
+CREATE TABLE "{target_schema}"."tmp_{target_table}"(
+    LIKE "{target_schema}"."{target_table}"
+)
