+++ projects/vdk-plugins/vdk-trino/src/vdk/plugin/trino/templates/load/dimension/scd1/05-insert-into-tmp-target.sql
@@ -1,4 +1,4 @@
-INSERT INTO "{target_schema}"."{target_table}"
+INSERT INTO "{target_schema}"."tmp_{target_table}"
 (
   SELECT *
   FROM   "{source_schema}"."{source_view}"
