+++ projects/vdk-plugins/vdk-trino/src/vdk/plugin/trino/templates/load/dimension/scd1/06-requisite-sql-scripts/06-drop-table.sql
