--- projects/vdk-plugins/vdk-trino/src/vdk/plugin/trino/templates/load/fact/insert/02-requisite-sql-scripts/02-insert-into-target.sql
@@ -1,5 +1,0 @@
-INSERT INTO {target_schema}.{target_table}
-(
-  SELECT *
-  FROM   {source_schema}.{source_view}
-)
