+++ projects/vdk-plugins/vdk-trino/src/vdk/plugin/trino/templates/load/fact/insert/06-requisite-sql-scripts/06-create-clone-table.sql
