+++ projects/vdk-plugins/vdk-trino/src/vdk/plugin/trino/templates/load/fact/insert/06-requisite-sql-scripts/06-create-consolidated-view.sql
@@ -1,11 +1,6 @@
 CREATE VIEW {view_schema}.{view_name}
 AS
-SELECT DISTINCT *
-FROM
 (
-    SELECT *
-    FROM {target_schema}.{target_table}
-    UNION ALL
     SELECT *
     FROM {staging_schema}.{staging_table}
 )
