+++ projects/vdk-plugins/vdk-trino/src/vdk/plugin/trino/templates/load/fact/snapshot/02-drop-backup-target.sql
@@ -1,0 +1,1 @@
+DROP TABLE IF EXISTS "{target_schema}"."backup_{target_table}"
