+++ projects/vdk-plugins/vdk-trino/src/vdk/plugin/trino/templates/load/fact/snapshot/03-drop-tmp-target.sql
@@ -1,0 +1,1 @@
+DROP TABLE IF EXISTS "{target_schema}"."tmp_{target_table}"
