+++ projects/vdk-plugins/vdk-trino/src/vdk/plugin/trino/templates/load/fact/snapshot/06-handle-quality-checks_and_move_data.py
@@ -11,7 +11,7 @@ from vdk.plugin.trino.trino_utils import TrinoTemplateQueries
 log = logging.getLogger(__name__)
 
 SQL_FILES_FOLDER = (
-    os.path.dirname(os.path.abspath(__file__)) + "/02-requisite-sql-scripts"
+    os.path.dirname(os.path.abspath(__file__)) + "/06-requisite-sql-scripts"
 )
 
 """
@@ -23,23 +23,22 @@ Otherwise the data will be directly processed according to the used template typ
 
 def run(job_input: IJobInput):
     """
-    1. Delete tmp_target and backup_target table if exist
-    2. create tmp_target
-    3. Insert source view data to tmp_target
-    4. if check,
-        - create staging table
-        - Use trino_utils function to move data from tmp_target to staging table
-        - send staging table for check validation
-        - If validated,
-            - drop backup table
-            - Use trino_utils function to move data from staging table to target table
-        - else Raise error
-    5. else,
-        - check if tmp_target table has data
-        - Use trino_utils function to move data from tmp target table to target table
+       1. Delete tmp_target and backup_target table if exist
+       2. create tmp_target
+       3. Insert target table and source view data to tmp_target using last_arrival_ts
+       4. if check,
+           - create staging table
+           - Use trino_utils function to move data from tmp_target to staging table
+           - send staging table for check validation
+           - If validated,
+               - drop backup table
+               - Use trino_utils function to move data from staging table to target table
+           - else Raise error
+       5. else,
+           - check if tmp_target table has data
+           - Use trino_utils function to move data from tmp target table to target table
 
     """
-
 
     job_arguments = job_input.get_arguments()
 
@@ -49,47 +48,17 @@ def run(job_input: IJobInput):
     target_schema = job_arguments.get("target_schema")
     target_table = job_arguments.get("target_table")
     trino_queries = TrinoTemplateQueries(job_input)
-    insert_query = CommonUtilities.get_file_content(SQL_FILES_FOLDER, "02-insert-into-target.sql")
-    drop_table_query = CommonUtilities.get_file_content(SQL_FILES_FOLDER, "02-drop-table.sql")
-    create_table_query = CommonUtilities.get_file_content(SQL_FILES_FOLDER, "02-create-table.sql")
+    drop_table_query = CommonUtilities.get_file_content(SQL_FILES_FOLDER, "06-drop-table.sql")
+    create_table_query = CommonUtilities.get_file_content(SQL_FILES_FOLDER, "06-create-table.sql")
 
     backup_target_table = f"backup_{target_table}"
-    # Drop backup target if already exists
-    drop_backup_target = drop_table_query.format(
-        target_schema=target_schema, target_table=backup_target_table
-    )
-    job_input.execute_query(drop_backup_target)
-
     tmp_target_table = f"tmp_{target_table}"
-    # Drop tmp target if already exists
-    drop_tmp_target = drop_table_query.format(
-        target_schema=target_schema, target_table=tmp_target_table
-    )
-    job_input.execute_query(drop_tmp_target)
 
-    #create tmp target table
-    create_staging_table = create_table_query.format(
-        table_schema=target_schema,
-        table_name=tmp_target_table,
-        target_schema=target_schema,
-        target_table=target_table
-    )
-    job_input.execute_query(create_staging_table)
-
-    # insert into tmp target table
-    insert_into_tmp = insert_query.format(
-        target_schema=target_schema,
-        target_table=tmp_target_table,
-        source_schema=source_schema,
-        source_view=source_view,
-    )
-    job_input.execute_query(insert_into_tmp)
-
+    # if validation check is selected
     if check:
         staging_schema = job_arguments.get("staging_schema", target_schema)
-        staging_table_name = CommonUtilities.get_staging_table_name(
-            target_schema, target_table
-        )
+        staging_table_name = CommonUtilities.get_staging_table_name(target_schema, target_table)
+
         staging_table = f"{staging_schema}.{staging_table_name}"
         target_table_full_name = f"{target_schema}.{target_table}"
 
@@ -112,28 +81,28 @@ def run(job_input: IJobInput):
 
         if check(staging_table):
             job_input.execute_query(f"SELECT * FROM information_schema.tables WHERE "
-                    f"table_schema = '{staging_schema}' AND table_name = '{staging_table_name}'")
+                        f"table_schema = '{staging_schema}' AND table_name = '{staging_table_name}'")
 
             # Drop backup target if already exists
             drop_backup_target = drop_table_query.format(
-                target_schema=target_schema, target_table=target_table
+                target_schema=target_schema, target_table=backup_target_table
             )
             job_input.execute_query(drop_backup_target)
 
             # use trino_utils function to move data
             trino_queries.perform_safe_move_data_to_table_step(
                 from_db=staging_schema,
-                from_table_name=staging_table,
+                from_table_name=staging_table_name,
                 to_db=target_schema,
                 to_table_name=target_table,
             )
 
         else:
-                raise DataQualityException(
-                    checked_object=staging_table,
-                    source_view=f"{source_schema}.{source_view}",
-                    target_table=target_table_full_name,
-                )
+            raise DataQualityException(
+                checked_object=staging_table,
+                source_view=f"{source_schema}.{source_view}",
+                target_table=target_table_full_name,
+            )
 
     else:
         log.debug("Check if tmp target has data.")
