+++ projects/vdk-plugins/vdk-trino/src/vdk/plugin/trino/templates/load/fact/snapshot/06-requisite-sql-scripts/06-create-table.sql
@@ -1,0 +1,4 @@
+-- create table with same schema as target table
+CREATE TABLE "{table_schema}"."{table_name}"(
+    LIKE "{target_schema}"."{target_table}"
+)
