+++ projects/vdk-plugins/vdk-trino/src/vdk/plugin/trino/templates/load/fact/snapshot/06-requisite-sql-scripts/06-drop-table.sql
@@ -1,0 +1,1 @@
+DROP TABLE IF EXISTS "{target_schema}"."{target_table}"
